Codechange: Split SettingEntry-related code to own file.
