+++ src/CMakeLists.txt
@@ -396,6 +396,8 @@ add_files(
     screenshot.h
     screenshot_type.h
     screenshot_pcx.cpp
+    settingentry_gui.cpp
+    settingentry_gui.h
     settings.cpp
     settings_cmd.h
     settings_func.h
